Now using a channel that ties back to Rust 1.65.0
There were issues building out on non ARM linux for me, that doesn't
happen in Github Actions for some reason, that was stopping local dev.
The crux of the issue was 1.65.0 was upgraded to 1.78.0 via the previous
channel config and that was missing a feature something was using to
compile, specifically it was

error[E0635]: unknown feature `proc_macro_span_shrink`
  --> /home/pfrank/.asdf/installs/rust/1.65.0/registry/src/index.crates.io-6f17d22bba15001f/proc-macro2-1.0.46/src/lib.rs:92:30
   |
92 |     feature(proc_macro_span, proc_macro_span_shrink)
   |                              ^^^^^^^^^^^^^^^^^^^^^^

This fixed that.